requirements.txt
torch>=2.0.0
transformers>=4.35.0
qwen-vl-utils==0.0.8
pillow>=10.0.0
matplotlib>=3.7.0
jupyter>=1.0.0
accelerate>=0.20.0
